+++ src/services/resources-getter.js
@@ -2,13 +2,12 @@ import _ from 'lodash';
 import P from 'bluebird';
 import { Schemas, logger } from 'forest-express';
 import Operators from '../utils/operators';
-import OperatorValueParser from './operator-value-parser';
 import CompositeKeysManager from './composite-keys-manager';
 import QueryBuilder from './query-builder';
 import SearchBuilder from './search-builder';
 import LiveQueryChecker from './live-query-checker';
 import { ErrorHTTP422 } from './errors';
-import Orm from '../utils/orm';
+import FiltersParser from './filters-parser';
 
 function ResourcesGetter(model, options, params) {
   const schema = Schemas.schemas[model.name];
@@ -17,28 +16,27 @@ function ResourcesGetter(model, options, params) {
   let segmentWhere;
   const OPERATORS = new Operators(options);
   const primaryKey = _.keys(model.primaryKeys)[0];
+  const filterParser = new FiltersParser(schema, params.timezone, options);
 
   function getFieldNamesRequested() {
     if (!params.fields || !params.fields[model.name]) { return null; }
 
     // NOTICE: Populate the necessary associations for filters
-    const associationsForQuery = [];
-    _.each(params.filter, (values, key) => {
-      if (key.indexOf(':') !== -1) {
-        const association = key.split(':')[0];
-        associationsForQuery.push(association);
-      }
-    });
+    const associations = params.filters ? filterParser.getAssociations(params.filters) : [];
 
-    if (params.sort && params.sort.indexOf('.') !== -1) {
-      associationsForQuery.push(params.sort.split('.')[0]);
+    if (params.sort && params.sort.includes('.')) {
+      let associationFromSorting = params.sort.split('.')[0];
+      if (associationFromSorting[0] === '-') {
+        associationFromSorting = associationFromSorting.substring(1);
+      }
+      associations.push(associationFromSorting);
     }
 
     // NOTICE: Force the primaryKey retrieval to store the records properly in the client.
     return _.union(
       [primaryKey],
       params.fields[model.name].split(','),
-      associationsForQuery,
+      associations,
     );
   }
 
@@ -53,28 +51,7 @@ function ResourcesGetter(model, options, params) {
   let hasSmartFieldSearch = false;
 
   function handleFilterParams() {
-    const where = {};
-    const conditions = [];
-
-    _.each(params.filter, (values, key) => {
-      if (key.indexOf(':') !== -1) {
-        const [associationName, fieldName] = key.split(':');
-        const columnName = Orm.getColumnNameForReferenceField(schema, associationName, fieldName);
-        key = `$${associationName}.${columnName}$`;
-      }
-      values.split(',').forEach((value) => {
-        const condition = {};
-        condition[key] = new OperatorValueParser(options)
-          .perform(model, key, value, params.timezone);
-        conditions.push(condition);
-      });
-    });
-
-    if (params.filterType) {
-      where[OPERATORS[params.filterType.toUpperCase()]] = conditions;
-    }
-
-    return where;
+    return filterParser.perform(params.filters);
   }
 
   function getWhere() {
@@ -86,7 +63,7 @@ function ResourcesGetter(model, options, params) {
         where[OPERATORS.AND].push(searchBuilder.perform());
       }
 
-      if (params.filter) {
+      if (params.filters) {
         where[OPERATORS.AND].push(handleFilterParams());
       }