+++ src/utils/orm.js
@@ -1,5 +1,3 @@
-import { Schemas } from 'forest-express';
-
 const semver = require('semver');
 
 const REGEX_VERSION = /(\d+\.)?(\d+\.)?(\*|\d+)/;
@@ -32,13 +30,6 @@ const getColumnName = (schema, fieldName) => {
   return (schemaField && schemaField.columnName) ? schemaField.columnName : fieldName;
 };
 
-const getColumnNameForReferenceField = (schema, associationName, fieldName) => {
-  const schemaField = schema.fields.find(field => field.field === associationName);
-  const [tableName] = schemaField.reference.split('.');
-  const associationSchema = Schemas.schemas[tableName];
-  return getColumnName(associationSchema, fieldName);
-};
-
 const isUUID = (DataTypes, fieldType) =>
   fieldType instanceof DataTypes.UUID
     || fieldType instanceof DataTypes.UUIDV1
@@ -48,5 +39,4 @@ exports.getVersion = getVersion;
 exports.isVersionLessThan4 = isVersionLessThan4;
 exports.findRecord = findRecord;
 exports.getColumnName = getColumnName;
-exports.getColumnNameForReferenceField = getColumnNameForReferenceField;
 exports.isUUID = isUUID;