+++ test/index.js
@@ -332,7 +332,7 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
               group_by_field: 'firstName',
               aggregate: 'Count',
               time_range: null,
-              filters: [],
+              filters: null,
             }, sequelizeOptions)
               .perform()
               .then((stat) => {
@@ -352,7 +352,7 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
               group_by_field: 'userAlias:id',
               aggregate: 'Count',
               time_range: null,
-              filters: [],
+              filters: null,
             }, sequelizeOptions)
               .perform()
               .then((stat) => {
@@ -375,7 +375,7 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
             group_by_date_field: 'createdAt',
             aggregate: 'Count',
             time_range: 'Day',
-            filters: [],
+            filters: null,
           }, sequelizeOptions)
             .perform()
             .then((stat) => {
@@ -395,7 +395,7 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
             group_by_date_field: 'createdAt',
             aggregate: 'Count',
             time_range: 'Week',
-            filters: [],
+            filters: null,
           }, sequelizeOptions)
             .perform()
             .then((stat) => {
@@ -415,7 +415,7 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
             group_by_date_field: 'createdAt',
             aggregate: 'Count',
             time_range: 'Month',
-            filters: [],
+            filters: null,
           }, sequelizeOptions)
             .perform()
             .then((stat) => {
@@ -435,7 +435,7 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
             group_by_date_field: 'createdAt',
             aggregate: 'Count',
             time_range: 'Year',
-            filters: [],
+            filters: null,
           }, sequelizeOptions)
             .perform()
             .then((stat) => {
@@ -726,12 +726,10 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
             user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken',
           },
           page: { number: '1', size: '30' },
-          filterType: 'and',
           timezone: 'Europe/Paris',
         };
 
         const paramsBaseCount = {
-          filterType: 'and',
           timezone: 'Europe/Paris',
         };
 
@@ -740,19 +738,21 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
             address: 'id,city,country',
           },
           page: { number: '1', size: '30' },
-          filterType: 'and',
           timezone: 'Europe/Paris',
         };
 
         const paramsAddressCount = {
-          filterType: 'and',
           timezone: 'Europe/Paris',
         };
 
         describe('with a "is" condition on a number field', () => {
           it('should return the records for the specified page', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { id: '100' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'equal',
+              value: 100,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -764,7 +764,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { id: '100' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'equal',
+              value: 100,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -778,7 +782,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "greater than" condition on a number field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { id: '>101' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'greater_than',
+              value: 101,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -790,7 +798,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { id: '>101' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'greater_than',
+              value: 101,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -804,7 +816,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "less than" condition on a number field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { id: '<104' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'less_than',
+              value: 104,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -816,7 +832,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { id: '<104' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'less_than',
+              value: 104,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -830,7 +850,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is not" condition on a number field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { id: '!100' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'not_equal',
+              value: 100,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -842,7 +866,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the records result', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { id: '!100' };
+            params.filters = JSON.stringify({
+              field: 'id',
+              operator: 'not_equal',
+              value: 100,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -856,7 +884,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is null" condition on a boolean field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { emailValid: 'null' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'equal',
+              value: null,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -868,7 +900,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { emailValid: 'null' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'equal',
+              value: null,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -882,7 +918,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is true" condition on a boolean field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { emailValid: 'true' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'equal',
+              value: true,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -894,7 +934,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { emailValid: 'true' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'equal',
+              value: true,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -908,7 +952,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is false" condition on a boolean field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { emailValid: 'false' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'equal',
+              value: false,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -920,7 +968,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { emailValid: 'false' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'equal',
+              value: false,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -934,7 +986,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is not null" condition on a boolean field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { emailValid: '!null' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'not_equal',
+              value: null,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -946,7 +1002,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { emailValid: '!null' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'not_equal',
+              value: null,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -960,7 +1020,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is not true" condition on a boolean field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { emailValid: '!true' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'not',
+              value: true,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -972,7 +1036,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { emailValid: '!true' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'not',
+              value: true,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -986,7 +1054,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is not" condition on a string field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { email: '!richard@piedpiper.com' };
+            params.filters = JSON.stringify({
+              field: 'email',
+              operator: 'not_equal',
+              value: 'richard@piedpiper.com',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -998,7 +1070,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { email: '!richard@piedpiper.com' };
+            params.filters = JSON.stringify({
+              field: 'email',
+              operator: 'not_equal',
+              value: 'richard@piedpiper.com',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1012,7 +1088,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is not false" condition on a boolean field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { emailValid: '!false' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'not',
+              value: false,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1024,7 +1104,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { emailValid: '!false' };
+            params.filters = JSON.stringify({
+              field: 'emailValid',
+              operator: 'not',
+              value: false,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1038,7 +1122,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "contains" condition on a string field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { firstName: '*Richa*' };
+            params.filters = JSON.stringify({
+              field: 'firstName',
+              operator: 'contains',
+              value: 'Richa',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1050,7 +1138,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { firstName: '*Richa*' };
+            params.filters = JSON.stringify({
+              field: 'firstName',
+              operator: 'contains',
+              value: 'Richa',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1064,7 +1156,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "not contains" condition on a string field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { username: '!*hello*' };
+            params.filters = JSON.stringify({
+              field: 'username',
+              operator: 'not_contains',
+              value: 'hello',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1076,7 +1172,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { username: '!*hello*' };
+            params.filters = JSON.stringify({
+              field: 'username',
+              operator: 'not_contains',
+              value: 'hello',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1090,7 +1190,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "starts with" condition on a string field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { email: 'dinesh@*' };
+            params.filters = JSON.stringify({
+              field: 'email',
+              operator: 'starts_with',
+              value: 'dinesh@',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1102,7 +1206,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { email: 'dinesh@*' };
+            params.filters = JSON.stringify({
+              field: 'email',
+              operator: 'starts_with',
+              value: 'dinesh@',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1116,7 +1224,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "ends with" condition on a string field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { email: '*@piedpiper.com' };
+            params.filters = JSON.stringify({
+              field: 'email',
+              operator: 'ends_with',
+              value: '@piedpiper.com',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1128,7 +1240,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { email: '*@piedpiper.com' };
+            params.filters = JSON.stringify({
+              field: 'email',
+              operator: 'ends_with',
+              value: '@piedpiper.com',
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1142,7 +1258,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is present" condition on a string field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsAddressList);
-            params.filter = { country: '$present' };
+            params.filters = JSON.stringify({
+              field: 'country',
+              operator: 'present',
+              value: null,
+            });
             new ResourcesGetter(models.address, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1158,7 +1278,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "is blank" condition on a string field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsAddressList);
-            params.filter = { country: '$blank' };
+            params.filters = JSON.stringify({
+              field: 'country',
+              operator: 'blank',
+              value: null,
+            });
             new ResourcesGetter(models.address, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1170,7 +1294,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsAddressCount);
-            params.filter = { country: '$blank' };
+            params.filters = JSON.stringify({
+              field: 'country',
+              operator: 'blank',
+              value: null,
+            });
             new ResourcesGetter(models.address, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1184,7 +1312,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "before x hours" condition on a date field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { createdAt: '$2HoursBefore' };
+            params.filters = JSON.stringify({
+              field: 'createdAt',
+              operator: 'before_x_hours_ago',
+              value: 2,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1196,7 +1328,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { createdAt: '$2HoursBefore' };
+            params.filters = JSON.stringify({
+              field: 'createdAt',
+              operator: 'before_x_hours_ago',
+              value: 2,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
@@ -1210,7 +1346,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
         describe('with a "after x hours" condition on a date field', () => {
           it('should generate a valid SQL query', (done) => {
             const params = _.clone(paramsBaseList);
-            params.filter = { createdAt: '$2HoursAfter' };
+            params.filters = JSON.stringify({
+              field: 'createdAt',
+              operator: 'after_x_hours_ago',
+              value: 2,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .perform()
               .then((result) => {
@@ -1222,11 +1362,64 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
           it('should return the total records count', (done) => {
             const params = _.clone(paramsBaseCount);
-            params.filter = { createdAt: '$2HoursAfter' };
+            params.filters = JSON.stringify({
+              field: 'createdAt',
+              operator: 'after_x_hours_ago',
+              value: 2,
+            });
             new ResourcesGetter(models.user, sequelizeOptions, params)
               .count()
               .then((count) => {
                 expect(count).equal(4);
+                done();
+              })
+              .catch(done);
+          });
+        });
+
+        describe('with complex filters conditions', () => {
+          const filters = JSON.stringify({
+            aggregator: 'and',
+            conditions: [{
+              field: 'createdAt',
+              operator: 'after_x_hours_ago',
+              value: 2,
+            }, {
+              aggregator: 'or',
+              conditions: [{
+                field: 'firstName',
+                operator: 'contains',
+                value: 'h',
+              }, {
+                field: 'lastName',
+                operator: 'starts_with',
+                value: 'Lumb',
+              }],
+            }, {
+              field: 'id',
+              operator: 'greater_than',
+              value: 12,
+            }],
+          });
+          it('should generate a valid SQL query', (done) => {
+            const params = _.clone(paramsBaseList);
+            params.filters = filters;
+            new ResourcesGetter(models.user, sequelizeOptions, params)
+              .perform()
+              .then((result) => {
+                expect(result[0]).to.have.length.of(3);
+                done();
+              })
+              .catch(done);
+          });
+
+          it('should return the total records count', (done) => {
+            const params = _.clone(paramsBaseCount);
+            params.filters = filters;
+            new ResourcesGetter(models.user, sequelizeOptions, params)
+              .count()
+              .then((count) => {
+                expect(count).equal(3);
                 done();
               })
               .catch(done);
@@ -1241,8 +1434,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
               user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken',
             },
             page: { number: '2', size: '50' },
-            filterType: 'and',
-            filter: { username: '*hello*' },
+            filters: JSON.stringify({
+              field: 'username',
+              operator: 'contains',
+              value: 'hello',
+            }),
             search: 'world',
             timezone: 'Europe/Paris',
           };
@@ -1257,8 +1453,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
         it('should return the total records count', (done) => {
           const params = {
-            filterType: 'and',
-            filter: { username: '*hello*' },
+            filters: JSON.stringify({
+              field: 'username',
+              operator: 'contains',
+              value: 'hello',
+            }),
             search: 'world',
             timezone: 'Europe/Paris',
           };
@@ -1316,8 +1515,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
               user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken',
             },
             page: { number: '2', size: '50' },
-            filterType: 'and',
-            filter: { username: '*hello*' },
+            filters: JSON.stringify({
+              field: 'username',
+              operator: 'contains',
+              value: 'hello',
+            }),
             sort: '-id',
             search: 'world',
             timezone: 'Europe/Paris',
@@ -1333,8 +1535,11 @@ const HasManyDissociator = require('../src/services/has-many-dissociator');
 
         it('should return the total records count', (done) => {
           const params = {
-            filterType: 'and',
-            filter: { username: '*hello*' },
+            filters: JSON.stringify({
+              field: 'username',
+              operator: 'contains',
+              value: 'hello',
+            }),
             search: 'world',
             timezone: 'Europe/Paris',
           };